Improve hook behavior for historical snapshots
Hook oid is not stored during startup anymore, it's loaded on demand in
the change callback and cached. The cache is invalidated by the syscache
invalidation hook. This ensures consistent and error-less behavior for
historical snapshot as well as correctly handling cases where function
oid was changed because of DROP/CREATE.

Note, we will continue replication even if the hook was dropped as the
existance of function is only checked at connect time.