Introduce a hash table based cache for hook functions
Instead of caching hook function information in the PGLogicalOutputData,
store it in a global hash table. Fixes memory errors in
hook_cache_callback found in valgrind.

We don't really nee to cache the function info beyond a single decoding
session. The problem is that we have to register a syscache invalidation
for the function, in case its definition changes during a session. That
syscache invalidation can get called after our decoding context ends if
we're called from the SQL interface. If it refers to the
PGLogicalOutputData after that decoding session has ended it'll access
freed memory.

Using a global hash in CacheMemoryContext ensures that the callbacks
always access valid memory.

We don't actually gain cross-decoding-session caching from this
because the SQL interface flushes the syscache on each call, but
it still gives us caching within the session.

This change fixes the registration of a separate syscache invalidation
callback each time decoding is started. Now it's registered only once
on startup.

Fixes #2