Implement the startup message
The protocol defines a startup message, which must be delivered once
before the first other message on the wire, and at no other time.

Implement it.

Currently it's deferred until the first transaction is decoded because
logical decoding doesn't let us enqueue messages on the wire during
init. Hopefully that can be improved in 9.6. In the mean time, clients
will have to force a dummy transaction after starting logical decoding
to get the startup message in a timely manner.