Fix walsender tests
Walsender based tests run again, and should now be closer to the
SQL decoding tests. Instead of trying to read and flush unwanted
changes in setUp, we now just delay creation of a slot until we
have done the setup and are ready to start the transactions for
the test its self.