Rework worker management
The worker management was split to separate module with own API that is
common for both db managers and apply workers. The storage now uses
array inside static shared memory because we need to allocate array of
max_worker_processes anyway so there is no point in having some workers
in static shared memory and some in dsm.