Replica identity is now checked in repset manipulation functions
Instead of checking for UPDATEs and DELETEs in the executor for every
command, check if the table has replica identity when it is added to
replication set which replicates updates and deletes. This performs much
better as it's one off check during administration procedure instead of
affecting every query.

This means that tables without primary key can't be in the default
replication set.

Administrator can obviously manipulate the catalog manually and break
the replication when we handle it like this, but that's not different
from any other catalog in PostgreSQL.