Implement C function based hook support
Rewrite the hook support to use a single C function in pg_proc
as an entry point, then plain C function pointers for the hooks.

Add an example that can be used to drive tests. It adapts the
C hooks to provide support for calling plpgsql or other SQL-level
functions. Include some dummy functions as examples, and allow
users to specify their own functions during setup.

This change removes the previous support for hooks defined in pg_proc
and invoked via the fmgr. The cached fmgr call infrastructure removed
by this change will be reusable in the pglogical_output_plhooks
extension to do faster calls, but for now is simply removed.