More replication set fixes
Cache invalidation now works correctly as we pass correct relation to
the api. Previously we passed the replication set relation which is
obviously wrong since the cache contains Oids of actual relations.

Also the replication set deletion deletes all the dependent records from
replication_set_table table.

One notable thing that is still missing is that we keep orphaned record
in replication_set_table when the table was dropped.