Install the exts for 'make check' in USE_PGXS mode
We can't do a pure "make check" because PGXS doesn't support making a
temp_install and we need one if we're going to install pglogical_output
and pglogical in order to run tests.

The prior 'make check' override rule assumed pglogical_output and
pglogical were installed. This meant you could be running 'make check'
against different code than what's in the current tree without
realising.

Instead, add pglogical_output as a submodule and install it before
installing pglogical, initdb'ing a temp instance and running the tests.

This isn't lovely, in that it modifies the PostgreSQL install tree,
which you wouldn't expect 'make check' to do. But at least it handles
all the config etc for you, unlike 'installcheck', and uses a temp
instance.

It may make sense to rename to 'tmpcheck' or 'instancecheck' or
something.