Enable pglogical_output caching support
Send the relmeta_cache_size parameter to the output plugin so that it
enables relation metadata caching instead of sending a new set of
relation metadata for each row.

This will dramatically improve performance for some workloads and
greatly reduce the memory churn on the downstream's relation metadata
cache.

Older versons of the output plugin don't understand the parameter and
ignore it, then continue sending metadata for each row. That's what
we already coped with before so no special action is needed.