Introduce version info for pglogical
Start at version 1.0.0. Make the extension script reflect that.

Generate the control file based on the version parsed out of
pglogical.h. This ensures the extension version always matches
the version in the code. That was a source of confusion with
BDR. It means we need dummy extension scripts for code-only
bumps, but that's pretty harmless and usefully explicit. We
can generate them with the Makefile if desired.

Add information functions to get the protocol versions and
pglogical version, and tests for them.

Introduce 'dist' and 'git-dist' make targets with support for
generating dist tarballs, checksums and detached signatures
for release management.