Use decoding context for relmetacache lifetime
We can actually scope the relation metadata cache to the decoding
session's lifetime. Even though the relcache invalidation callbacks
outlive it they can just check and see that the hash table pointer is
null so there's nothing to do right now.

That way we get a clean start each session and we always purge the
table at the end of the session.