Better management of subscription changes
This commit and the previous one uncoveder fundamental locking issues in
the way we handle nodes. For quick and dirty hack we protect all the
manipulation interfaces using exclusive lock on local node table (as it
only has one record which is for local node, so we can lock the table
itself), but longterm better locking will need to be done. That will
probably mean rewriting at least part of the pglogical_node api.