Shuffle header files
Create two new headers, for definitions of some struct that are causing
a complicated tangle of inclusions.  Only typedefs to structs appear in
widely included headers, which removes any circularity without need for
any forward struct declarations.  The definitions are hidden in
"internal" files that are only included by the .c files that need them.
This also reduces the exposure of .c files to additional Postgres
includes.

The overall structure could use some more cleanup, but this seems a
useful step forward.