Fix linking
The 4afdadf added pglogical_create_subscriber.o to OBJS so that it's
cleaned on make clean. The unintended consequence of that was that it
also got linked into pglogical.so extension which is not desirable. Move
it to EXTRA_CLEAN instead which achieves the make clean removal goal
without other side effects.