Fix a race condition between managers and short-lived workers
There was a race where a manager might register a bgworker for launch then not
get rescheduled for some time. During this window its worker might launch,
attach to shmem, do its work, detach cleanly and release the slot. Another
manager can then pick the slot up and register a different worker on it. When
the first manager gets scheduled and wakes up it checks to see if the worker it
launched started, sees that it's in BGW_STOPPED state, and assumes it crashed
before successfully attaching to shmem, so it sets the crashed_at field.

Because crashed_at is set, another manager can then pick up the slot and
register another worker on it. So we have at least one abandoned worker which
might be reading from / writing to the same slot as another worker, getting
data not meant for it, and generally making a mess. The problem cascades too,
since the abandoned worker will release the slot when it's launched even though
it's now claimed by another worker, causing the problem to repeat.

To fix this introduce a generation counter for worker slots. A manager can now
tell the difference between a crashed worker and one that did its work quickly
before the manager was rescheduled by examining the generation field.

Crashed before attach: generation is the same as at registration so we know
to set crashed_at on behalf of the crashed worker.

Launched and attached then did work and exited cleanly / crashed, slot re-used:
generation has been incremented by the registration of a new worker in this
slot; we know the worker attached and don't have to take action ourselves.

This commit also adds some DEBUG2 logging around worker management to make any
future diagnostic work easier. Some assertions are also added.

See RT#42356 for details and history.