Fix ERROR on commit timestamp lookup of frozen xid
Pg emits an ERROR if you try to pass FrozenTransactionId (2) or
BootstrapTransactionId (1) to TransactionIdGetCommitTsData, per RT#46983 . This
seems like an oversight in the core function, but we can work around it here by
setting it to the same thing we'd get if the xid's commit timestamp was trimmed
already.

This fixes

    ERROR: cannot retrieve commit timestamp for transaction 2

errors on apply.