Merge pglogical_output into pglogical.
We still build pglogical_output library, but only to load the pglogical
and we do this only for backwards compatibility. For same reason we
still create slot with pglogical_output. We can change this in next
major version.

This gets rid of hooks and Makefile magic to make pglogical_output work.

TODO:
Protocol specific tests.
Cleanup internals documentation.