Completely rewrite the TAP tests
A new minimal framework is introduced: a subclass PostgresNode,
PostgresPGLNode, and a class for an individual pglogical "node" (database),
PGLDB. The naming is awkward because PostgresNode uses "node" to mean a Pg
instance and we use "node" in pglogical to mean a pglogical-enabled DB.

The pglogical_create_subscriber test has been completely rewritten to use
the new framework. So has the non-defaul replication sets test.

The remaining test, 030_pglogical_daylight_times_switching.pl, is disabled
for now, and has been tasked-out for a rewrite.

This work will make it practical to write useful tests in future that work
on a variety of hosts as part of routine test runs.

commit 81a19344a890daee3e9af04f798d857d9df9369a
Author: Craig Ringer <craig@2ndquadrant.com>
Date:   Wed May 31 13:17:30 2017 +0800

    Allow TAP to find modules

commit 92418b23c7ffc50346dca6fa23caeb56720a0f8d
Author: Craig Ringer <craig@2ndquadrant.com>
Date:   Wed May 31 13:17:06 2017 +0800

    Typo: could [not] fetch