Improve error message when a DEFERRABLE constraint is downstream
If there's a DEFERRABLE constraint on the downstream pglogical cannot currently
apply an insert/update because it doesn't know how to do index rechecks. The error
it emitted was:

    ERROR: pglogical doesn't support index rechecks

which is arguably less than helpful. Improve the error to include an
explanation, the affected index(es) and the base relation.