Correctly compute current_timestamp on downstream
All uses of current_timestamp were getting the same value as when it was first
invoked, even across transaction boundaries.

StartTransactionCommand() calls StartTransaction() in xact.c, which sets the
initial timestamp. But it doesn't clear it; instead it sets it to
stmtStartTimestamp, which is not part of the transaction state stack. In a
normal backend, PostgresMain calls SetCurrentStatementStartTimestamp() before
query execution, so it's outside the normal transaction stack management path
pglogical uses.

pglogical does not execute statements, as such, so it doesn't call
SetCurrentStatementStartTimestamp() at any point.

To fix this, treat each applied transaction as a "statement" for this purpose,
since really we only care about the transaction timestamp, and call
SetCurrentStatementStartTimestamp() before we StartTransactionCommand()
during apply.

Fixes 2ndQuadrant/pglogical_internal#148