Prefer the replica identity index when handling insert conflict checks
Check the replica identity index, if any, when handling checks for
insert conflicts. This should deliver a more consistent and predictable
result during apply.

Also remove the unreachable test for multiple conflicts. We used to scan the
remaining indexes to see if we found a second conflict and ERROR there, but
some time before this code moved over from BDR we stopped doing that. Instead
we let the normal executor handle the unique violation ERROR at apply-time.