Disregard partial unique indexes to prevent false conflicts
An unique index like

    UNIQUE (a) WHERE (b)

cannot safely be used for conflict detection/resolution. If there's
also a subscriber tuple 'a=x' satisfying 'b' and the remote inserts
a new tuple with 'a=x' that does not satify b, the remote insert
will succeed. The tuple doesn't violate the index's uniqueness.

On the downstream we used to consider this index for conflict
resolution, so we'd look up 'a=x' in the index and find the existing
local tuple. We didn't test the new tuple's 'b', so we determined
there was a conflict, and replaced the existing local tuple.

This caused divergence and false conflicts.

As an immediate workaround we can just disregard partial unique
indexes.

It would be better to ask the executor to evaluate the index predicate
against the incoming tuple. That way we could do conflict resolution
when actually warranted.

Works around RM#1839