Ensure process_syncing_tables returns to MessageContext
process_syncing_tables could switch to TopMemoryContext unintetionally. This
was probably OK because we'd reset the context when we next re-entered the
message handling loop, but posed some risk of leaks especially since it did
multiple rounds of work within multiple transactions. Be more careful to return
to MessageContext after each one.