Support re-creating replication slot used for table sync if it exists
If we crash during table sync we might have a pre-existing slot.
Prior to this change, we'd ERROR with an error like

    FATAL: could not send replication command "CREATE_REPLICATION_SLOT "pgl_dbname_subname_slotname_sync" LOGICAL pglogical_output": status PGRES_FATAL_ERROR: ERROR: replication slot "pgl_dbname_subname_slotname_sync" already exists

We now detect this and drop and re-create the slot instead.

This could be further refined by skipping the initial COPY and skipping
straight to catchup if we detect that a prior sync attempt copied the table but
failed during catchup. But making that safe may require further changes to
ensure that we remember the stop-lsn for this sync in the catalogs, otherwise
the new sync run would set a later stop-lsn. It'd also require us to restart
the worker on failure, which we don't do yet.

With this change we'll at least restart table syncs after a nodec crash/restart
properly.

This is a backport of pglogical3 functionality.