Tweak pglogical resync test for concurrency
Make sure that the pglogical resync concurrency test runs
pgbench during the resync of the tables, and that it adds the
tables to the repset before it starts pgbench.

The idea is to capture evidence of a suspected bug where a resync process can
fail with a walsender timeout as seen in RT#60889.

This only fails at larger scales / smaller timeouts.