Allow tables to be resynchronized in state SYNC_STATUS_SYNCDONE
Currently a table cannot be resynchronized when in SYNC_STATUS_SYNCDONE.
If a relation has been resynchronized, it will remain in state
SYNC_STATUS_SYNCDONE until the LSN has advanced.  A subscriber node may
run a DML statement but if an attempt to resynchronize the relation
while still in state SYNC_STATUS_SYNCDONE, then an error will be thrown
because the relation is "is already being synchronized".