Document and update test for synchronize_structure as text
This is to allow distinct values for this subscription parameter.
I've choosed to update default existing tests.

This commit also add sql for the new synchronize_structure (as text)

Default to "none" instead of "false". Possible values are:
 * none (was 'false)
 * all (was 'true')

Filenames with an X because I dont' know what the next version will be.