synchronize structure only of replicated relations
It's using the -t switch from pg_dump, thus with the same limitations.
Also using 'strict-names' now in order to make the pg_dump filtering
more robust.

At the moment we track only sequences and tables in pglogical, so this new
feature only dump/restore tables and linked objects (index, constraint,
sequences). Materialized view, for example, are not dump/restore at the moment.

More logic can be added for this feature to be complete.

This is dropping the old function create_subscription() using a boolean: because of
the default values for parameters, PostgreSQL cannot decide what
function to execute when 2 have the same non default parameters.

This should be explicitely written in the Release Notes to help users
upgrading their scripts.

It is required to build a list of replicated seqeunce, not only tables
for RPC calls