Only import replication origin headers after compat loads
We need HAVE_REPLICATION_ORIGINS to be defined from pglogical.h,
but it needs to be included late because it defines macros that
cause problems interpreting Pg headers.

Really we should split compat??/pglogical_compat.h into a pre-
and post- set, with the post-set being all the function wrapper
macros.