Add in_repliate_ddl_command variable
Add a global in_repliate_ddl_command variable that mirrors the one in
pgl3, which may be used in user ProcessUtility_hook code to test whether
it's running in the context of pglogical.replicate_ddl_command.

This is not exposed to SQL because it's not much use. The event triggers
for sql_drop and ddl_command_end fire after pglogical_replicate_ddl_command
has returned. There's no way to annotate the information they collect and
return with whether they were replicated or not. Also, for drops, the catalog
changes are already visible so we can't even see if the table was replicated
or not.

RM#5881