Update protocol of write_rel()
This is to allow target name of a replicated table (or sequence) to differ from the local
name of the table (or sequence).

New reltarget and nsptarget parameters are used as input instead of the Relation
information. This is because the target names can differ from the local names
(we are on the provider) when using write_rel).

For instance:
    data->api->write_rel(ctx->out, data, relation, att_list,
                         nsptarget, reltarget);

Add 2 new columns to replication_set_*

Attribute "nsptarget" and attribute "reltarget" or "seqtarget", according to the
"replication_set_*" table name.

Update output plugin to decode change

Update RPC functions to use the nsptarget/reltarget,..

Use nsptarget/reltarget for TRUNCATE
 because we send information for an object and all its replication set, we now
 compute a message for each unique (reloid,nsptarget,reltarget) triplet

Add the functions and tests to offer table mapping feature

From this commit the target can be distinct from the origin table/sequence.

Add regression test for the table mapping feature

Try to outline usual expected usage of the feature and some corner
cases.

Add documentation related to reltarget and nsptarget
This include the OUTPUT internal document.

Prevent multiple origin from the same set to the same target

Allow to return multiple source for a table
  SQL function (show repset tables by target)