Add a new "force_text_transfer" subscription option
Trying to replicate a text column on the provider into a uuid column on
the subscriber had strange effects (this was never advertised to work in
pglogical, but it works with native logical replication). For example,
the text value '8a829417540af4a101541021883b12d0' was not replicated as
the UUID value '8a829417-540a-f4a1-0154-1021883b12d0' as expected, but
the strange value '43386138-3239-3431-3735-343061663461' instead (i.e.,
not 0x8a, 0x82, … but the literal bytes '8', 'a', '8', '2', … from the
varlena representation).

This change adds a new subscription option, "force_text_transfer", that
sets binary.want_internal_basetypes and binary.want_binary_basetypes to
0 in pglogical_start_replication(), to request the provider to send the
text representations of all data types. This is a very big hammer which
turns off the usual type-specific decision making altogether.

Using text for everything has a performance impact, so it should be used
only when necessary, for example to perform a zero-downtime type change
through a pglogical migration.