make {Acquire,Release}DeletionLock() private
Functions of the same name exist in core PostgreSQL. Previously, PostgreSQL's
versions were not visible to the outside world, but 5be594caf81 moved
those function's declaration to a public header file and removed the
static flag on their declaration. That commit was backpatched into 9.6.18,
10.13, 11.8 and 12.3. Unfortunately, this clashes with pglogical's own
copies of those functions. To avoid #if-labyrinths, just rename the
functions in pglogical.