Fix dump/restore execution
Correctly escape the connection string for both pg_dump and
pg_restore so that exotic database and user names are handled
correctly.

Use execv rather than system to do the actual execution, so
that the command line is built by the OS rather than us. We
do this to prevent potential escape to the shell and which
could let user who can create subscriptions execute
arbitrary shell commands as the user under which the PostgreSQL
instance is running.

Reported by Pedro Gallegos

Fixes BDR-384