Get rid of pgl_replorigin_drop() (mostly)
In PostgreSQL 10, replorigin_drop() got a nowait=false option for use
in DROP SUBSCRIPTION.  Because in pglogical 2, we used
pgl_replorigin_drop(), this option was not exposed, and so we couldn't
use this, even though we should.

So get rid of pgl_replorigin_drop() and call replorigin_drop()
directly with the options appropriate for each call.  For older
PostgreSQL versions, we ignore the nowait option.  That possibly
exposes the user to race conditions that the nowait=false option was
supposed to fix, but that's the best we can do, and in any case it
matches the current behavior.

I kept pgl_replorigin_drop() as the name for the custom implementation
for PostgreSQL 9.4, to keep that a bit separate.

This refactoring is also useful with an eye toward PostgreSQL 14,
where replorigin_drop() was removed and we will further need to
refactor this code.  Might as well get it to a clean state before we
do that.