Fix pglogical_show_subscription_table() result protocol (#110)
The implementation behaved like a set-returning function in
materialize mode, but it was not actually declared as a set-returning
function.  Apparently, this worked so far but it stopped working in
PostgreSQL 14 (see PostgreSQL commit
d4545dc19b8ea670bf62e06d22b0e4e6fcb45153).  So rewrite it as a normal
function returning a composite type.