Handle command line quoting for child procs under Windows
The exec_cmd function introduced by 95c0e898 used the _spawnv function
from the MS VC runtime to invoke the child process on Windows.

Unfortunately this function does NOT handle process arguments like
execv. Instead of preserving the argument vector, it concatenates it
with spaces between entries. No attention whatosever is paid to quoting
and escaping, so the result is garbage if there are spaces, double
quotes, etc in the argument(s).

Of course the _spawnv docs don't mention this, or warn that it's not an
analogue of exevc like you might anticipate. They don't link to the
docs for CommandLineToArgV[1] where the MSVCRT's conventions for decoding
command line arguments are documented either.

_spawnv also has a 1024 character limit for the command line, which
could well become an issue with pg_dump command invocations, long
connstrs with options and GUCs, etc.

Windows does not appear to have anything like execv. Nor does it have
any inverse of the CommandLineToArgV runtime function to help developers
correctly convert an argument vector into a string suitable for passing
to CreateProcess[2] or similar. You're expected to roll your own.

See the 2011 blog "Everyone quotes command line arguments the wrong
way"[3] for the incredibly gory details. Apparently nothing has improved
in the situation since then.

An additional problem with exec_cmd() using _spawn is that it does not
allow postgres to CHECK_FOR_INTERRUPTS() while it runs. So the backend
will be unresponsive to shutdown requests, pg_terminate_backend(), etc.

Accordingly, this patch changes exec_cmd to prepare a properly escaped
single argument string suitable for input to the child process's
CommandLineToArgV(). Then it uses the win32 API functions
CreateProcess(...), WaitForSingleHandle(...), GetExitCodeProcess(...)
etc to launch the process and wait for it to exit while continuing
to process interrupts.

Note that we will need a separate patch on non-win32 to either handle an
EINTR return from waitpid(), CHECK_FOR_INTERUPTS() and resume waiting,
or to waitpid() with a timeout that polls for interrupts.

[1] https://docs.microsoft.com/en-us/windows/win32/api/shellapi/nf-shellapi-commandlinetoargvw
[2] https://docs.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-createprocessa
[3] https://docs.microsoft.com/en-gb/archive/blogs/twistylittlepassagesallalike/everyone-quotes-command-line-arguments-the-wrong-way