Make sure we use a correctly deformed tuple when doing initial sync.
Formerly, we just used the original retrieved tuple in
pglogical_table_data_filtered() to initially sync a table. This
caused us to probably miss not yet materialized tuple column
values, like it happens if a column value is not yet materialized
when having pg_attribute.atthasmissing set after adding a column
with a non-null DEFAULT expression since PG11.

To fix this, deform the tuple and use a proper new version with
all column values replaced before filtering and storing it
into the tuplestore.