Fix building against Postgres 9.6
The root cause here appears to be that `ReadDirExtended` in Postgres
core is defined as `static` up until PG10.2. At that point it's
available publicly.

There seems to be some confusion on previous issues reported on this
repo: in #227 @petere says "You need at least PostgreSQL 10.2." --
presumably in context he must mean "to use 10.x you need at least 10.2"
since the compat9{4,5,6} directories clearly imply the intent to support
versions prior to 10. Later in #275 @petere says "pglogical 2.3.2 builds
correct against PostgreSQL 9.5." -- but I don't see how this can be the
case given this function is used but not available in the public API.

Here I've copied the function directly from Postgres core (but without
the static declaration) to support the usage in this project.