Fix resource leaks
A few functions failed to release resources in early returns. I
noticed the one with row filters while working on snapshot management
changes in PostgreSQL when I added an assertion that when a historic
snapshot is free'd, it must not still be marked as active. The leaks
in row filter processing are otherwise harmlesss AFAICS, but if a lot
of rows are filtered out, the resource owner holding all the resources
can grow very large.

The other leak is with multi-inserts to a table. I'm not sure what the
effect of that is, I found it by searching for other uses of
PushActiveSnapshot.